@@ -1,10 +1,10 @@
 using System.Linq;
 using System.Net.Mime;
 using System.Reflection;
+using System.Text.Json;
 using Microsoft.AspNetCore.Diagnostics.HealthChecks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
-using Newtonsoft.Json;
 
 namespace street_service.Utils.HealthChecks
 {
@@ -34,7 +34,7 @@ namespace street_service.Utils.HealthChecks
 
         private static string ProcessUnhealthy(HealthReport report)
         {
-            return JsonConvert.SerializeObject(new
+            return JsonSerializer.Serialize(new
                 {
                     application = new {
                         name = _assembly.Name,
@@ -55,7 +55,7 @@ namespace street_service.Utils.HealthChecks
 
         private static string ProcessHealthy(HealthReport report)
         {
-            return JsonConvert.SerializeObject(new
+            return JsonSerializer.Serialize(new
                 {
                     application = new {
                         name = _assembly.Name,
