@@ -8,10 +8,12 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using StockportGovUK.AspNetCore.Middleware;
 using StockportGovUK.AspNetCore.Availability;
+using StockportGovUK.AspNetCore.Availability.Middleware;
 using StockportGovUK.NetStandard.Gateways;
 using Swashbuckle.AspNetCore.Swagger;
 using street_service.Providers;
 using street_service.Services;
+using street_service.Utils.ServiceCollectionExtensions;
 
 namespace street_service
 {
@@ -27,35 +29,18 @@ namespace street_service
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddSingleton<IStreetProvider, FakeStreetProvider>();
-            services.AddSingleton<IStreetProvider, VerintStreetProvider>();
-            services.AddSingleton<IStreetService, StreetService>();
-
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddResilientHttpClients<IGateway, Gateway>(Configuration);
+            services.AddAvailability();
+            services.AddSwagger();
             services.AddHealthChecks()
                 .AddCheck<TestHealthCheck>("TestHealthCheck");
-            services.AddAvailability();
-            services.AddResilientHttpClients<IGateway, Gateway>(Configuration);
-            services.AddSwaggerGen(c =>
-            {
-                c.SwaggerDoc("v1", new Info { Title = "street_service API", Version = "v1" });
-                c.AddSecurityDefinition("Bearer", new ApiKeyScheme
-                {
-                    Description = "Authorization using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
-                    Name = "Authorization",
-                    In = "header",
-                    Type = "apiKey"
-                });
-                c.AddSecurityRequirement(new Dictionary<string, IEnumerable<string>>
-                {
-                    {"Bearer", new string[] { }},
-                });
-            });
+
+            services.RegisterServices();
         }
 
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)
         {
-            if (env.IsDevelopment())
+            if (env.IsEnvironment("local"))
             {
                 app.UseDeveloperExceptionPage();
             }
@@ -63,16 +48,21 @@ namespace street_service
             {
                 app.UseHsts();
             }
-            app.UseMiddleware<ExceptionHandling>();
-            app.UseHttpsRedirection();
-            app.UseHealthChecks("/healthcheck", HealthCheckConfig.Options);
-            app.UseMvc();
-            app.UseSwagger();
 
-            var swaggerPrefix = env.EnvironmentName == "local" ? string.Empty : "/streetservice";
+            app.UseHttpsRedirection();
+            app.UseRouting();
+            app.UseEndpoints(endpoints => endpoints.MapControllers());
+
+            app.UseMiddleware<Availability>();
+            app.UseMiddleware<ApiExceptionHandling>();
+            
+            app.UseHealthChecks("/healthcheck", HealthCheckConfig.Options);
+
+            app.UseSwagger();
             app.UseSwaggerUI(c =>
             {
-                c.SwaggerEndpoint($"{swaggerPrefix}/swagger/v1/swagger.json", "street_service API");
+                c.SwaggerEndpoint(
+                    $"{(env.IsEnvironment("local") ? string.Empty : "/streetservice")}/swagger/v1/swagger.json", "Street service API");
             });
         }
     }
